refactor: 패키지 구조 변경 (support에서 handler, http, status로)
