+++ tomcat/src/main/java/org/apache/coyote/handler/StaticHandlerMethod.java
@@ -1,6 +1,13 @@
-package org.apache.coyote.support;
+package org.apache.coyote.handler;
 
 import org.apache.constant.MediaType;
+import org.apache.coyote.http.HttpHeader;
+import org.apache.coyote.http.HttpRequest;
+import org.apache.coyote.http.HttpResponse;
+import org.apache.coyote.http.HttpStatus;
+import org.apache.coyote.status.HttpCookie;
+import org.apache.coyote.status.Session;
+import org.apache.coyote.status.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import support.IoUtils;
