+++ tomcat/src/main/java/org/apache/coyote/http/HttpParser.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.support;
+package org.apache.coyote.http;
 
 import static java.util.Arrays.stream;
