+++ tomcat/src/main/java/org/apache/coyote/status/Session.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.support;
+package org.apache.coyote.status;
 
 import java.util.HashMap;
 import java.util.Map;
